Use our own TTF version of Courier font by default
On some systems the Type 1 Courier font as shipped by texlive [1] will
be used as the default display font. For unknown reasons this doesn't
render well in Bless on some systems.

This commit adds our own TTF version of the aforementioned Courier font,
to ensure we have a nice looking default Courier font even on systems
that don't ship one, or on which the (Type 1) Courier font is problematic.

The font name of bless-courier.ttf is 'BlessCourier', to allow using the
'Courier' font name for system-wide Courier font if needed.

[1] https://www.tug.org/texlive/Contents/live/texmf-dist/fonts/type1/adobe/courier/pcrr8a.pfb

Fixes #25